feat: add single-file DailyAffirmationDisplay implementing ProgramInterface
